+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Fix PDOStatement deprecation notice on `bindParam()` (#586)
+- Fix deprecation notice thrown when instrumenting the `PDOStatement::bindParam()` method and passing `$length = null` (#586)
 
 ## 4.2.6 (2022-01-10)