+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- ...
+
+## 4.2.7 (2022-02-18)
+
 - Fix deprecation notice thrown when instrumenting the `PDOStatement::bindParam()` method and passing `$length = null` (#586)
 
 ## 4.2.6 (2022-01-10)